app/requirements.txt
gradio==3.16.2
torch>=1.10
torchvision>=0.11
matplotlib
numpy
Pillow
python-dotenv
google-genai>=0.0.9
